created test view and can successfully log in and sign up
