+++ views/login.handlebars
@@ -1,5 +1,5 @@
 <h2 class="page-header">Account Login</h2>
-<form method="post" action="/login">
+<form method="post" action="/users/login">
   <div class="form-group">
     <label>Username</label>
     <input type="text" class="form-control" name="username" placeholder="Username">
