+++ views/register.handlebars
@@ -10,7 +10,7 @@
         <div class="alert alert-danger">{{msg}}</div>
       {{/each}}
 {{/if}}
-<form method="post" action="/register">
+<form method="post" action="/users/register">
    <div class="form-group">
     <label>Name</label>
     <input type="text" class="form-control" placeholder="Name" name="name">
